# docs/api/openapi.yaml
info:
  description: 'Human-facing REST API for QuantDinger. Agent integrations use `/api/agent/v1`
    — see `docs/agent/agent-openapi.json`. Conventions: `docs/API_CONVENTIONS.md`.'
  contact:
    name: QuantDinger
    url: https://github.com/brokermr810/quantdinger
  license:
    name: See repository LICENSE
  x-api-app-version: 3.0.22
  title: QuantDinger Web API
  version: 1.0.0
components:
  securitySchemes:
    HumanJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'User session JWT from `POST /api/auth/login`. Send as `Authorization:
        Bearer <token>`.'
    AgentToken:
      type: http
      scheme: bearer
      bearerFormat: qd_agent_xxx
      description: Agent Gateway token for `/api/agent/v1/*` only. See docs/agent/AGENT_QUICKSTART.md.
  schemas:
...
    HumanSuccessEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 1 on success
          example: 1
        msg:
          type: string
          example: success
        data:
          nullable: true
      additionalProperties: false
    HumanErrorEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 0 on error
          example: 0
        msg:
          type: string
        data:
          nullable: true
      additionalProperties: false
...
servers:
- url: http://localhost:5000
  description: Backend direct (python run.py)
- url: http://localhost:8888
  description: Docker Compose (via reverse proxy)
paths:
...
  /api/community/indicators:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: List published marketplace indicators.
      description: "Query params:\npage: Page number (default 1)\npage_size: Page\
        \ size (default 12)\nkeyword: Search keyword\npricing_type: 'free' / 'paid'\
        \ / empty (all)\nsort_by: 'score' (default) / 'newest' / 'hot' / 'price_asc'\
        \ /\n         'price_desc' / 'rating'.\n         'score' sorts by the composite\
        \ multi-factor backtest score\n         (see services/experiment/scoring.py)\
        \ and is now the\n         default — putting genuinely well-performing indicators\n\
        \         at the top of the marketplace, not just the newest ones."
      tags:
      - Community
      operationId: getApiCommunityIndicators
      x-visibility: public
  /api/community/indicators/{indicator_id}:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get marketplace indicator detail.
      tags:
      - Community
      operationId: getApiCommunityIndicatorsIndicatorId
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/community/indicators/{indicator_id}/purchase:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Purchase a marketplace indicator.
      description: 'Side effects:

        1. Verify buyer has enough credits

        2. Debit buyer credits and credit seller

        3. Create purchase record

        4. Copy indicator into buyer account'
      tags:
      - Community
      operationId: postApiCommunityIndicatorsIndicatorIdPurchase
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/community/indicators/{indicator_id}/sync:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Sync latest code for a purchased indicator.
      description: "Use when the publisher updated code after the buyer purchased;\
        \ the buyer\npulls the newest version into their local copy.\n\nPreconditions:\n\
        \    - Caller must have purchased the indicator\n    - Original listing must\
        \ still be published"
      tags:
      - Community
      operationId: postApiCommunityIndicatorsIndicatorIdSync
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/community/my-purchases:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: List indicators purchased by the current user.
      tags:
      - Community
      operationId: getApiCommunityMyPurchases
      x-visibility: public
  /api/community/author/summary:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: 'Author overview: published / approved / pending / sales / revenue
        / avg rating.'
      tags:
      - Community
      operationId: getApiCommunityAuthorSummary
      x-visibility: public
  /api/community/author/published:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get Api Community Author Published
      tags:
      - Community
      operationId: getApiCommunityAuthorPublished
      description: List indicators published by the author with per-item sales, revenue,
        and rating.
      x-visibility: public
  /api/community/author/sales:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Paginated author sales ledger; optional indicator_id filter.
      tags:
      - Community
      operationId: getApiCommunityAuthorSales
      x-visibility: public
  /api/community/indicators/{indicator_id}/comments:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: List comments for an indicator.
      tags:
      - Community
      operationId: getApiCommunityIndicatorsIndicatorIdComments
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Add a comment.
      description: "Request body:\n    rating: 1-5 star rating\n    content: Optional\
        \ comment text (max 500 chars)\n\nOnly purchasers may comment, and only once\
        \ per indicator."
      tags:
      - Community
      operationId: postApiCommunityIndicatorsIndicatorIdComments
      x-visibility: public
  /api/community/indicators/{indicator_id}/comments/{comment_id}:
    put:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Update own comment.
      description: 'Request body:

        rating: 1-5 star rating

        content: Comment text (max 500 chars)'
      tags:
      - Community
      operationId: putApiCommunityIndicatorsIndicatorIdCommentsCommentId
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
    - in: path
      name: comment_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/community/indicators/{indicator_id}/my-comment:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get current user's comment on an indicator (for edit form).
      tags:
      - Community
      operationId: getApiCommunityIndicatorsIndicatorIdMyComment
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/community/indicators/{indicator_id}/performance:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get live performance statistics for an indicator.
      tags:
      - Community
      operationId: getApiCommunityIndicatorsIndicatorIdPerformance
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/community/admin/pending-indicators:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: List indicators pending review (admin only).
      description: 'Query params:

        page: Page number (default 1)

        page_size: Page size (default 20)

        review_status: ''pending'' / ''approved'' / ''rejected'' / ''all'''
      tags:
      - Community
      operationId: getApiCommunityAdminPendingIndicators
      x-visibility: public
  /api/community/admin/review-stats:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Review queue statistics (admin only).
      tags:
      - Community
      operationId: getApiCommunityAdminReviewStats
      x-visibility: public
  /api/community/admin/indicators/{indicator_id}/review:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Approve or reject an indicator (admin only).
      description: 'Request body:

        action: ''approve'' / ''reject''

        note: Optional review note'
      tags:
      - Community
      operationId: postApiCommunityAdminIndicatorsIndicatorIdReview
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/community/admin/indicators/{indicator_id}/unpublish:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Unpublish an indicator (admin only).
      description: 'Request body:

        note: Optional unpublish reason'
      tags:
      - Community
      operationId: postApiCommunityAdminIndicatorsIndicatorIdUnpublish
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/community/admin/indicators/{indicator_id}:
    delete:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Delete an indicator (admin only).
      tags:
      - Community
      operationId: deleteApiCommunityAdminIndicatorsIndicatorId
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
...
openapi: 3.0.3